refactor(repo): finish the purity wall — inject tz into analytics, move to @intake/core (phase 3.1)
Completes Phase 3's deferred "purity wall". correlateTimeSeries was the only
genuinely wall-clock-impure analytics function (it bucketed days via local
getters and parsed an unzoned date string). It now takes an injected
`timezone: string = "UTC"` and is deterministic: day-bucketing uses
`Intl.DateTimeFormat("en-CA", { timeZone })` and the lag-shift uses UTC
calendar arithmetic. With the host-zone dependency gone, all of
analytics-stats.ts moves into the pure, I/O-free @intake/core
(@intake/core/analytics-stats); the analytics types/consts move to
@intake/types/analytics (PhaseType sourced from @intake/types/records).

The sole production caller (analytics-service.correlate) passes
getDeviceTimezone() so user-facing correlations anchor to the viewer's zone.
buildInsightsPrompt is left untouched in apps/web — its `new
Date(numberTs).toISOString()` is already UTC-deterministic; it is AI-prompt-
coupled, so it rides with the future @intake/ai-prompts (Phase 4), not the
wall-clock work.

Re-export shims at @/lib/analytics-stats and @/lib/analytics-types keep all
33 importers unchanged (zero churn).

Locks the wall shut: add packages/core/eslint.config.mjs — a standalone,
purity-only flat config (deliberately NOT spreading @intake/eslint-config's
strict base) that bans Date.now(), bare new Date(), Math.random() and fetch()
via no-restricted-syntax (new Date(arg) stays allowed). `turbo run lint` /
CI's `pnpm lint` pick up the new @intake/core lint task automatically.

Verified: turbo typecheck (4 pkgs) + lint (0 errors; the purity-lint catches
Date.now/new Date()/Math.random in a negative test), 2023 unit tests (incl. 2
new tz-bucketing cases proving UTC vs America/New_York), 33 integration
(testcontainers), production build, bundle-scan. 3-agent adversarial review
(correctness / boundary / CI) — no blockers; the tz/lag math is boundary-safe
and DST-immune. DEFERRED.md updated (wall-clock + purity-lint items struck;
analytics-insights split re-scoped to Phase 4).
