+++ packages/core/src/index.ts
@@ -8,10 +8,13 @@
  * PURITY GUARD: this package's tsconfig deliberately ships NO DOM lib and NO
  * `@types/node`, so `window` / `document` / `localStorage` / `btoa` / `atob` /
  * `navigator` / `process` are not in scope — referencing them is a compile
- * error. (Wall-clock readers — `Date.now()` / `new Date()` — are still in the
- * ES2022 lib, so they are NOT structurally blocked; core code must instead take
- * an injected `now`. The two date-dependent analytics functions stay in
- * apps/web until the Phase 3.1 now/tz-injection follow-up.)
+ * error. Wall-clock readers (`Date.now()` / bare `new Date()`) and `Math.random`
+ * are still in the ES2022 lib, so they are additionally banned by this package's
+ * `eslint.config.mjs` (`no-restricted-syntax`) — core takes an injected `now`/
+ * `tz` instead. (Example: `correlateTimeSeries` takes a `timezone` param rather
+ * than reading the host zone. `buildInsightsPrompt` deliberately stays in
+ * apps/web — it is already pure but is AI-prompt-coupled, so it rides with the
+ * future `@intake/ai-prompts` extraction.)
  *
  * Subpath exports mirror the file layout — prefer the granular paths
  * (`@intake/core/compound`, `@intake/core/security`, …) over this barrel.
@@ -23,3 +26,4 @@ export * from "./settings-helpers";
 export * from "./service-result";
 export * from "./security";
 export * from "./shake-detector";
+export * from "./analytics-stats";
